Remove compatibility code
We were using `six` to write Python 2 and 3 compatible code. With the
recent changes around the Python versions we support, it is not
necessary anymore.

This PR removes all usages of `six` with their Python 3 equivalents.

Also, removes the compatibility code written by us, such as defining
things that are not in Python 2.

Finally, this PR contains very minor changes to some parts to
get rid of some warnings.