Type hint the public API
This PR introduces type hints for the public API we provide.
Note that, we haven't started type hinting the internals, as we
plan to do it incrementally.

There is no mypy setup associated with this PR, as it would require
more changes, which can be done later.

Apart from type hints, we also fixed some documentation problems
and tried to follow 80 character documentation length throughout
the client code base.