Convert `Config` classes to public API
These were converted into private classes during the 4.0 migration,
as we were only supporting keyword arguments for the client configuration.

However, that led to a poor user experience, as we had to use `**kwargs`,
instead of listing all the configuration elements in the client constructor.
(IDEs become painfully slow once we list all keyword arguments, as there
are too many of them)

The solution to this problem is to make the Config classes public API
and make the client able to use it directly.

```python
config = Config()
config.cluster_name = "dev2"

client = HazelcastClient(config)
```

We provide full type hints for config elements.