Make the tests more resilient to timing problems
We had some test failures on the backward compatibility runs, mainly
due to the fact that, the failing tests were not resilient enough
to timing problems.

In the `HeartbeatTest::test_heartbeat_stopped_and_restored`, it was
possible that the responses coming from the server for the previous
heartbeat requests to override what we wrote as the `last_read_time`.

In the `ListenerAddMemberTest::test_add_member`, it was possible that
the `map#put` would complete before the listener is registered to the
newly added connection.