Implement compact serialization
This PR introduces new compact serialization to the Python client.

Users interact with this API through, `compact_serializers` configuration
element, which will be documented along with other aspects of the
implementation in the future PRs.

`compact_serializers` is a dictionary from the type of user class to
`CompactSerializer` associated with it.

Schema registration and fetching mechanisms are implemented with
a minimum overhead to other serialization options.

They follow the structures described as `Controlled Serialization`
and `Eager Deserialization` in the metadata distribution design.

In this PR, only `map#put` and `map#get` APIs support compact
serialization. Other APIs will be added in further PRs to not
complicate this already huge PR.