[API-1048] Make intellisense and type hints available on blocking proxies (#519)
* Make intellisense and type hints available on blocking proxies

This PR attempts to provide type hints and intellisense for the
blocking proxies by introducing blocking versions of the proxies
that implement the same interface with different return types.

With this approach, we don't have to duplicate the documentation,
and return the blocking subclasses directly to the user.

Actually, the blocking proxies are not really subclasses, but very
thin wrappers around the public APIs of the proxies. They look like
subclasses to make the documentation of the actual proxies available
to them.

This makes mypy complain, but it is possible to ignore such
warnings about return type mismatches between child and parent classes.
So, we could work around that problem once we start using mypy.

Also, we have modified the return types and documentation of transactional
proxies. It wasn't possible to use them in non-blocking mode so I changed
the return types and documentation to reflect that.

* convert timeout to int for default values of the semaphore#try_acquire