[API-521] Add support for compact serialization to public APIs (#531)
* Add support for compact serialization to public APIs

For compact serialization to work, we have to control the results
of the serialization. Also, for the responses we got from the server,
we have to eagerly deserialize them, and retry after receiving the
schema from the cluster.

With this PR, we make sure that on each API we support compact serialization,
we check the result of the `to_data` call to see if it throws or not.

If it throws, we try to send the schema to the cluster and retry.

Also, this PR adds tests for all the APIs we support compact serialization.
For now, we didn't implement eager deserialization for the APIs that
return `ImmutableLazyDataList`, so there are no tests for such APIs.

To be able to add those tests, we would need quite a bit of compact
serializable classes written in Java. This PR includes aforementioned
Java code, but they might be moved to the remote controller package so
that it would be accessible to all clients.

* address review comments

* assign values to correct field names in the TopicMessage constructor

* improve sql compatibility test

* address review comments