Fix mypy errors
There were around 60 errors shown in the mypy checks about the various
minor problems across the client code base. This PR corrects
those errors and adds a check for the mypy to PR builder workflow.

Note that, for now, we are only checking the type hints in the
hazelcast directory (client's code base). Tests and examples are not
passing the mypy checks for now. We can improve this incrementally,
and add them in the future.

Also, we are not using the strict flag of the mypy, due to not
type hinting the internals yet. Hopefully, we will have that
in the future.