Eagerly deserialize SQL columns of type OBJECT
Due to lazy deserialization, Compact serialized objects were
not working nicely with SQL. It was possible to get deserialization
errors that cannot be easily understood by the user.

Similar to what we did for other APIs, we have removed lazy
deserialization from this API as well.

To do that efficiently, I have modified the protocol code generator
so that we can pass to_object function to the SqlPageCodec where
we can do deserialization while filling rows and columns, so that
we will not iterate over the result set once more.