Eagerly deserialize responses consisting of list of data [API-1964] (#618)
* Eagerly deserialize responses consisting of list of data

We were performing lazy deserialization on APIs that return
a list of data, such  as map#values(). This was not going to work
with Compact, as after returning the list to the user, there might
be Compact serialized data on the list which we don't have the
schema on the client yet. If it was a normal response, the client
would have fetched the schema, but it is not possible with
these lazy APIs. We perform the deserialization while getting
the list items and this is a sync API. We can't perform a blocking
call there, because there is a chance that this will be executed
in the reactor thread, which would result in a deadlock.

So, the only possible way of making these APIs work with Compact
is to convert them to eager deserialization.

This PR removes lazy deserialization in everything apart from SQL,
which will be handled in another PR.

* fix mypy error

* address review comments