Update Let's Encrypt certificate periodically
This PR adds a workflow to update the Let's Encrypt certificate
used in tests periodically.

The workflow updates the certificate used in tests every month,
based on the certificate generated in the private test artifacts
repo. The certificate is renewed every week there, and each certificate
is valid for 3 months, so renewing the certificate in this repo
every month should work without a problem.