Remove Sphinx version duplication (#695)
The version specified in the documentation is provided by the Sphinx
config in `docs/conf.py`.

This duplicates the version specified in `hazelcast/init.py`, and
previously [became
outdated](https://github.com/hazelcast/hazelcast-python-client/pull/690).

Unfortunately, in this PR I updated the version incorrectly (i.e. to
`5.4.0`, not `5.5.0`), highlighting the issues with duplicating this
data.

Instead, we should have a single place the version if specified, and
reference it in the Sphinx config.

Example output following this change:
<img width="898" alt="image"
src="https://github.com/user-attachments/assets/36718c1f-fdfc-406c-bbf8-31d9ac207ff5">

In addition, while trying to build the documents locally, `make html`
failed with a cryptic error message:
> *** commands commence before first target. Stop.

I didn't have Sphinx installed, and while `make` was happy to evaluate
the `if` check for it's existence, printing an error message was
disallowed outside of a target. Fixed by adding a target, and requiring
that on any Sphinx related targets.

This can be tested by updating the `SPHINXBUILD` command to something
that doesn't exist and trying to build the documentation:
> % make html
Makefile:13: *** The 'non-existant-sphinx-build' command was not found.
Make sure you have Sphinx installed, then set the SPHINXBUILD
environment variable to point to the full path of the
'non-existant-sphinx-build' executable. Alternatively you can add the
directory with the executable to your PATH. If you don\'t have Sphinx
installed, grab it from http://sphinx-doc.org/. Stop.