Client compatibility repo updates (#763)
Hazelcast 5.4.4 and below doesn't work with the
`com.hazelcast.nio.ssl.BasicSSLContextFactory` to specify SSL options.
So, I changed it to use
`com.hazelcast.nio.ssl.ClasspathSSLContextFactory`
That requires embedding various certificates used by the Python Client
in the `hazelcast-enteprise-X-tests.jar`.

This PR also uses RCD as the test runner, since that's what we use on
the compatibiliy repo, and it makes easy to use the modified enterprise
test jar.

Once this PR is merged, I'll update the Python compatibility workflow
with similar changes.

NOTE: The jobs in this PR fail since the tests run with the old
workflow.
See
https://github.com/hazelcast/hazelcast-python-client/actions/runs/20167632094
for a passing run with the updates in this PR.